Change implementation of Reset function in VP8 wrapper
The Reset function was modified so that the encoder is destroyed
and recreated on reset. Initialization of the encoder and setting
of the encoder speed is now done in a private method, to avoid
code duplication. (It is used both in InitEncode and in Reset.)
This change is needed to make the unit tests pass with newer
versions of libvpx.
Review URL: http://webrtc-codereview.appspot.com/33004

git-svn-id: https://webrtc.googlecode.com/svn/trunk@56 4adac7df-926f-26a2-2b94-8c16560cd09d