Proof-of-concept proposal for a standalone webrtc build without using gyp_chromium etc. This adds the necessary scripts and gyp files. The idea is to assume that we are building within Chromium; in that case common.gypi (which every gyp file includes) provides the necessary logic to build webrtc.
In a standalone build, gyp_webrtc would be called, which includes common_standalone.gypi. This file specifies everything that running gyp_chromium would normally provide (cflags etc). Here we can customize things for our build that Chromium might not have, and also do away with a lot of the complexity which we don't need.

Most of the remaining work would be in common_standalone.gypi to provide full build settings. Much of this could come from Chromium's common.gypi.

Some of the inspiration for this is from NaCl.

(This doesn't impact the current build, just provides the option to run gyp_webrtc instead of gyp_chromium).
Review URL: http://webrtc-codereview.appspot.com/22021

git-svn-id: https://webrtc.googlecode.com/svn/trunk@60 4adac7df-926f-26a2-2b94-8c16560cd09d